api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.8
fonttools==4.56.0
h11==0.14.0
holidays==0.66
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prophet==1.1.6
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sniffio==1.3.1
stanio==0.5.1
starlette==0.45.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
